+++ ui/src/app/core/manage-plugins/manage-plugins.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 
 import { ManagePluginsModalComponent } from './manage-plugins-modal/manage-plugins-modal.component';
+import { UninstallPluginsModalComponent } from './uninstall-plugins-modal/uninstall-plugins-modal.component';
 import { SettingsPluginsModalComponent } from './settings-plugins-modal/settings-plugins-modal.component';
 import { CustomPluginsService } from './custom-plugins/custom-plugins.service';
 
@@ -24,12 +25,12 @@ export class ManagePluginsService {
     ref.componentInstance.pluginName = pluginName;
   }
 
-  uninstallPlugin(pluginName) {
-    const ref = this.modalService.open(ManagePluginsModalComponent, {
-      size: 'lg',
+  uninstallPlugin(pluginName, settingsSchema) {
+    const ref = this.modalService.open(UninstallPluginsModalComponent, {
       backdrop: 'static',
     });
     ref.componentInstance.action = 'Uninstall';
+    ref.componentInstance.settingsSchema = settingsSchema;
     ref.componentInstance.pluginName = pluginName;
   }